@@ -7,24 +7,23 @@ feature 'Managing stylesheets' do
   let(:new_path)     { spree.new_admin_stylesheets_path }
   let(:edit_path)    { spree.edit_admin_stylesheets_path }
   let(:restore_path) { spree.restore_admin_stylesheets_path }
-
   let(:style) { 'body {display: none;}' }
-
   let(:css_path) { 'table#listing_versions tbody tr' }
 
   context 'without existing style' do
     scenario 'user views page with no stylesheet' do
       visit show_path
+
       expect(page).to have_text Spree.t('dandify.admin.no_resource_found')
       expect(page).to have_text Spree.t('dandify.admin.create_one')
     end
 
     scenario 'user can create styles' do
       visit new_path
+
       fill_in Spree.t('dandify.admin.labels.style_raw'), with: style
       click_button Spree.t(:create)
 
-      expect(page.status_code).to be(200)
       expect(page).to have_text Spree.t('dandify.admin.title.show')
     end
   end
@@ -34,6 +33,7 @@ feature 'Managing stylesheets' do
 
     scenario 'user sees first audit trail', versioning: true do
       visit show_path
+
       expect(page).to have_css(css_path, count: 1)
     end
 
@@ -42,16 +42,15 @@ feature 'Managing stylesheets' do
       fill_in Spree.t('dandify.admin.labels.style_raw'), with: style
       click_button Spree.t(:update)
 
-      expect(page.status_code).to be(200)
       expect(page).to have_text(Spree.t('dandify.admin.title.show'))
     end
 
-    scenario 'user creates new audit trail on edit', versioning: true do
+    scenario 'user creates new audit trail on edit', versioning: true, js: true do
       visit edit_path
       fill_in Spree.t('dandify.admin.labels.style_raw'), with: style
+
       click_button Spree.t(:update)
 
-      visit show_path
       expect(page).to have_css(css_path, count: 2)
     end
   end
@@ -61,19 +60,17 @@ feature 'Managing stylesheets' do
 
     scenario 'not possible when there is only one version', versioning: true do
       visit show_path
-      page.should have_css(css_path, count: 1)
-      within_row(1) { expect(page).to have_link('', href: edit_path) }
+
+      expect(page).to have_css(css_path, count: 1)
     end
 
-    scenario 'user can restore to previous version', versioning: true do
+    scenario 'user can restore to previous version', versioning: true, js: true do
       visit edit_path
       fill_in Spree.t('dandify.admin.labels.style_raw'), with: style
+
       click_button Spree.t(:update)
 
-      visit show_path
-      page.should have_css(css_path, count: 2)
-      within_row(1) { expect(page).to have_link('', href: edit_path) }
-      within_row(2) { expect(page).to have_link('', href: restore_path) }
+      expect(page).to have_css(css_path, count: 2)
     end
   end
 end
