@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe 'Managing stylesheet' do
+feature 'Managing stylesheet' do
   stub_authorization!
 
   let(:show_path) { spree.admin_stylesheets_path }
@@ -10,6 +10,7 @@ describe 'Managing stylesheet' do
   context 'create stylesheet link' do
     it 'is available when there are no stylesheets' do
       visit show_path
+
       expect(page).to have_link(
         Spree.t('dandify.admin.buttons.new'), href: new_path
       )
@@ -17,7 +18,9 @@ describe 'Managing stylesheet' do
 
     it 'is not available when there are stylesheets' do
       create(:stylesheet)
+
       visit show_path
+
       expect(page).to_not have_link(
         Spree.t('dandify.admin.buttons.new'), href: new_path
       )
@@ -27,6 +30,7 @@ describe 'Managing stylesheet' do
   context 'edit stylesheet link' do
     it 'is not available when there are stylesheets' do
       visit show_path
+
       expect(page).to_not have_link(
         Spree.t('dandify.admin.buttons.edit'), href: edit_path
       )
@@ -34,33 +38,32 @@ describe 'Managing stylesheet' do
 
     it 'is available when there are stylesheets' do
       create(:stylesheet)
+
       visit show_path
+
       expect(page).to have_link(
         Spree.t('dandify.admin.buttons.edit'), href: edit_path
       )
     end
   end
 
-  context 'back to list of stylesheets button' do
+  context 'cancel button' do
     it 'is available on `new` page' do
       visit new_path
-      expect(page).to have_link(
-        Spree.t('dandify.admin.buttons.back'), href: show_path
-      )
+
+      expect(page).to have_link(Spree.t('actions.cancel'), href: show_path)
     end
 
     it 'is available on `edit` page' do
       visit edit_path
-      expect(page).to have_link(
-        Spree.t('dandify.admin.buttons.back'), href: show_path
-      )
+
+      expect(page).to have_link(Spree.t('actions.cancel'), href: show_path)
     end
 
     it 'is not available on listing page' do
       visit show_path
-      expect(page).to_not have_link(
-        Spree.t('dandify.admin.buttons.back'), href: show_path
-      )
+
+      expect(page).not_to have_link(Spree.t('actions.cancel'), href: show_path)
     end
   end
 end
