@@ -10,30 +10,33 @@ rescue LoadError
   exit
 end
 
+require 'capybara'
+require 'capybara/poltergeist'
+require 'capybara/rails'
+require 'capybara/rspec'
+require 'database_cleaner'
+require 'ffaker'
+require 'paper_trail/frameworks/rspec'
+require 'pry-byebug'
 require 'rspec/rails'
 require 'shoulda-matchers'
-require 'ffaker'
-require 'database_cleaner'
-require 'capybara'
-require 'capybara/rspec'
-require 'capybara/rails'
-require 'capybara/poltergeist'
-require 'paper_trail/frameworks/rspec'
 
 Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
 
-require 'spree/testing_support/factories'
-require 'spree/testing_support/controller_requests'
 require 'spree/testing_support/authorization_helpers'
-require 'spree/testing_support/url_helpers'
 require 'spree/testing_support/capybara_ext'
+require 'spree/testing_support/controller_requests'
+require 'spree/testing_support/factories'
+require 'spree/testing_support/url_helpers'
+
+Capybara.javascript_driver = :poltergeist
 
 FactoryGirl.find_definitions
 
 RSpec.configure do |config|
   config.deprecation_stream = 'rspec.log'
-  config.include Spree::TestingSupport::ControllerRequests
   config.include FactoryGirl::Syntax::Methods
+  config.include Spree::TestingSupport::ControllerRequests, type: :controller
   config.include Spree::TestingSupport::UrlHelpers
   config.include Devise::TestHelpers, type: :controller
   config.use_transactional_fixtures = false
@@ -54,6 +57,4 @@ RSpec.configure do |config|
   config.after do
     DatabaseCleaner.clean
   end
-
-  Capybara.javascript_driver = :poltergeist
 end
